Consolidated CPU HISQ stencil routines into a single function.
