+++ tests/llfat_reference.h
@@ -10,6 +10,14 @@ extern "C"{
 			  void** ghost_sitelink_diag, QudaPrecision prec, void* act_path_coeff);
 
   void computeLongLinkCPU(void** longlink, void **sitelink, QudaPrecision prec, void* act_path_coeff);  
+
+  // CPU-style BLAS routines
+  void cpu_axy(QudaPrecision prec, double a, void* x, void* y, int size);
+  void cpu_xpy(QudaPrecision prec, void* x, void* y, int size);
+
+  void computeHISQLinksCPU(void** fatlink, void** longlink, void** fatlink_eps, void** longlink_eps,
+        void** sitelink, void* qudaGaugeParamPtr, 
+        double** act_path_coeffs, double eps_naik);
 #ifdef __cplusplus
 }
 #endif
